Changes for SFT contracts (#6)
* [SOL] Reverse dependency relationship between parent classes

This satisfies the following two goals:

* Superclasses should be specified in order of most basic to most derived
* Superclasses should be initialized in order of most basic to most derived

The order of subclasses entails a depends-on relationship between later
classes and earlier classes. We should be careful that our private
interfaces don't unnecessarily have large dependencies like ERC1155, so
we list private interfaces before the heavier dependencies.

* [SOL] Improve consistency of category separators

When adding new categories, a problem I noticed was that counting
"=" signs was difficult and led to inconsistent justification
(distance from the left of the IDE).

My solution is to left-align category names with a normal // comment
for consistent justification.

Separation is then depicted vertically with lines of comments, which
are easily copy-pastable for consistent length, and in leiu of
copy-paste, the author can hold "/" until their IDE counts past
character 80.

* [SOL] Organize functions to reflect dependency relationships

Just like our parent class reorg, we go from most basic to most derived.
The interface is the only thing external code sees, so all function
documentation should live there.

This commit is focused on organization, not rewriting documentation,
so documentation on overrides is just pasted next to documentation
in the interface. A later commit will address this.

* [SOL] De-duplicate and reword interface comments

The previous commit moved function doc without modifying the text,
leading to duplication where both interfaces and overrides were
documented.

This commit de-duplicates, and also rewords with my best understanding
of the interface.

* [SOL] Disambiguate "id"

I _think_ "id" universally refers to token IDs, but better to
disambiguate and remove all doubt.