[SOL] Consistency in function parameter names
My argument for removing "_" preceding variable names is that "_" is used
to indicate restricted access, such as for private member variables.

Function parameters do not have restricted access because they have full
function scope, so it seems better to have consistency on the side of no
preceding "_".

No logic is changed.