[SOL] Improve security hygiene
The general pattern of re-entrancy attacks is to attack where a state
update occurs after a call to an external contract, instead of before.

I *think* there isn't a vulnerability here, because the external
contract is trusted in this case (deployed by us).

However, good security hygiene is to proceed in the following order,
from most trusted to least trusted:

  * Validations
  * Access state
  * Update state
  * Calls to parent contracts
  * Calls to trusted external contracts
  * Calls to untrusted external contracts

A good example is TradeFloor.sol:

https://github.com/wolvesofwallstreet/wolves.finance/blob/c9c6d88/contracts/src/token/TradeFloor.sol

There, I use patterning (consistent placement of logic across the file)
to draw visibility to the need to scrutinize re-entrancy hygiene.